+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.1.4 (unreleased)
 
 - Added many loss functions
+- Added distance functions
 
 ## 0.1.3 (2019-11-30)